build: Move LDFLAGS to appear after objects
Linkers operate in order, with later objects on the command-line
providing definitions of symbols that are undefined in earlier objects.
As a result they expect to see dependent objects first, followed by
their dependencies.

Historically, this has been avoided by over-linking: with -lm in LDFLAGS,
the object was given an (at the time) unnecessary dependency on libm,
which later becomes necessary when the object files are added.

However, to reduce unnecessary dependencies and simplify the dependency
graph, some distributions' compilers (including Ubuntu and Fedora)
default to linking with the equivalent of -Wl,--as-needed, which defeats
that. As a result, now that we link with -lm, the build fails on those
linkers unless we move -lm to appear after the objects that depend on it.

Signed-off-by: Simon McVittie <smcv@debian.org>